feat: enable presets in multibot configuration
