+++ common-lib/src/main/kotlin/util/DeveloperOptions.kt
@@ -1,0 +1,7 @@
+package com.github.heheteam.commonlib.util
+
+import com.github.heheteam.commonlib.Student
+
+data class DeveloperOptions(
+  val presetStudent: Student? = null,
+)
