+++ multi-bot/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     implementation(project(":parent-bot"))
 
     implementation(libs.tgbotapi)
+    implementation(libs.kotlin.result)
     implementation(libs.exposed.spring.boot.starter)
     implementation("com.github.ajalt.clikt:clikt:5.0.2")
 }
